Move config options to a config.h that is not uploaded to git
config.h.template is the example to use, copy this and name it config.h in the include folder.  Edit the variables there and then each of us can have our own config thats specific to our device/location without affecting the main code.

If you add to config.h MAKE SURE YOU ADD to config.h.template as well so we know what needs to be defined in config.h